cloud functions/requirements.txt
pandas
pyarrow
numpy
google-cloud-storage
google-cloud-bigquery
google-cloud-functions
fsspec
gcsfs